chore(send queue): move code for media upload to its own file
